profile activity compose topappbar added,login activity cleanup
