+++ app/src/main/AndroidManifest.xml
@@ -26,6 +26,7 @@
         tools:targetApi="31">
         <activity
             android:name=".presentation.features.userprofile.ProfileActivity"
+            android:theme="@style/Theme.AppCompat.NoActionBar"
             android:exported="false" />
         <activity
             android:name="com.thefinestartist.finestwebview.FinestWebViewActivity"
