+++ app/src/main/res/layout/activity_profile.xml
@@ -6,15 +6,28 @@
     android:layout_height="match_parent"
     tools:context=".presentation.features.userprofile.ProfileActivity">
 
+    <androidx.compose.ui.platform.ComposeView
+        android:id="@+id/profileTopAppBar"
+        android:layout_width="0dp"
+        android:layout_height="?attr/actionBarSize"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintBottom_toTopOf="@id/profile_tab_layout"/>
+
+
     <com.google.android.material.tabs.TabLayout
         android:id="@+id/profile_tab_layout"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:backgroundTint="@color/dark_gray"
+        android:background="@color/dark_gray"
+        app:layout_constraintTop_toBottomOf="@id/profileTopAppBar"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintBottom_toTopOf="@id/profile_pager"
         app:tabTextColor="@color/white"
         app:tabIndicatorColor="@color/white"
-        app:tabMode="auto"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:tabMode="auto" />
 
     <androidx.viewpager2.widget.ViewPager2
         android:id="@+id/profile_pager"
@@ -22,6 +35,5 @@
         android:layout_height="0dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/profile_tab_layout" />
-
 
 </androidx.constraintlayout.widget.ConstraintLayout>
